wrote parser for advanced select() options
Working on a complimentary function for select()
that will handle more complicated arguments (and
then we can get rid of split_select() and
numeral() which both kind of suck). The actual
argument passed in is very simple to write, in
the form 'command arg1 arg2:type'. This update
implements a parser for that format to turn it
into something easier to use when matching
input.