cancelling gmalt; it was a bad idea
Originally, I was very excited to have a
graphical wrapper for malt. The plan was to allow
non-console users simply launch the program from
a desktop shortcut and not have to worry about
compatibility or installation or whatever. Turns
out that was not a great idea. You can do that
with console programs anyway. So gmalt is done.