rewrote serve() and deprecated lots of options
serve() has been reworked to be much cleaner
and to handle types that have caused trouble in
the past. It uses hasattr() instead of type() in
an attempt to keep things fairly flexible.
Removed lists of keywords; every command now has
exactly one keyword associated with it. Replaced
a bunch of global options. Planning on removing
old select() codes, but holding on to a stub for
BACK_CODE for right now for compatibility.