Separating back and front logic.
The problem I've run into here is that I want to
have a more advanced frontend interface for
complicated projects, but malt messes everything
up because it prints errors by itself. We're going
to solve this by removing the print statements
from backend logic, and adding more helpful params
to the response object, for the user to use how
they will.